cluster/spec: add test cases for cross cluster conflict check
