+++ pkg/cluster/spec/spec_test.go
@@ -114,261 +114,6 @@ pd_servers:
 	c.Assert(topo.PDServers[1].DataDir, Equals, "/test-data/pd-12379")
 }
 
-func (s *metaSuiteTopo) TestDirectoryConflicts1(c *C) {
-	topo := Specification{}
-
-	err := yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: "test-deploy"
-  data_dir: "test-data"
-tidb_servers:
-  - host: 172.16.5.138
-    deploy_dir: "/test-1"
-pd_servers:
-  - host: 172.16.5.138
-    data_dir: "/test-1"
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "directory conflict for '/test-1' between 'tidb_servers:172.16.5.138.deploy_dir' and 'pd_servers:172.16.5.138.data_dir'")
-
-	err = yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: "test-deploy"
-  data_dir: "/test-data"
-tikv_servers:
-  - host: 172.16.5.138
-    data_dir: "test-1"
-pd_servers:
-  - host: 172.16.5.138
-    data_dir: "test-1"
-`), &topo)
-	c.Assert(err, IsNil)
-
-	// report conflict if a non-import node use same dir as an imported one
-	err = yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: deploy
-  data_dir: data
-tidb_servers:
-  - host: 172.16.4.190
-    deploy_dir: /home/tidb/deploy
-pd_servers:
-  - host: 172.16.4.190
-    imported: true
-    name: pd_ip-172-16-4-190
-    deploy_dir: /home/tidb/deploy
-    data_dir: /home/tidb/deploy/data.pd
-    log_dir: /home/tidb/deploy/log
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "directory conflict for '/home/tidb/deploy' between 'tidb_servers:172.16.4.190.deploy_dir' and 'pd_servers:172.16.4.190.deploy_dir'")
-
-	// two imported tidb pass the validation, two pd servers (only one is imported) don't
-	err = yaml.Unmarshal([]byte(`
-global:
-  user: "test2"
-  ssh_port: 220
-  deploy_dir: deploy
-  data_dir: /data
-tidb_servers:
-  - host: 172.16.4.190
-    imported: true
-    port: 3306
-    deploy_dir: /home/tidb/deploy1
-  - host: 172.16.4.190
-    imported: true
-    status_port: 3307
-    deploy_dir: /home/tidb/deploy1
-pd_servers:
-  - host: 172.16.4.190
-    imported: true
-    name: pd_ip-172-16-4-190
-    deploy_dir: /home/tidb/deploy
-  - host: 172.16.4.190
-    name: pd_ip-172-16-4-190-2
-    client_port: 2381
-    peer_port: 2382
-    deploy_dir: /home/tidb/deploy
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "directory conflict for '/home/tidb/deploy' between 'pd_servers:172.16.4.190.deploy_dir' and 'pd_servers:172.16.4.190.deploy_dir'")
-}
-
-func (s *metaSuiteTopo) TestPortConflicts(c *C) {
-	topo := Specification{}
-	err := yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: "test-deploy"
-  data_dir: "test-data" 
-tidb_servers:
-  - host: 172.16.5.138
-    port: 1234
-tikv_servers:
-  - host: 172.16.5.138
-    status_port: 1234
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "port conflict for '1234' between 'tidb_servers:172.16.5.138.port' and 'tikv_servers:172.16.5.138.status_port'")
-
-	topo = Specification{}
-	// tispark_masters has "omitempty" in its tag value
-	err = yaml.Unmarshal([]byte(`
-monitored:
-  node_exporter_port: 1234
-tispark_masters:
-  - host: 172.16.5.138
-    port: 1234
-tikv_servers:
-  - host: 172.16.5.138
-    status_port: 2345
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "port conflict for '1234' between 'tispark_masters:172.16.5.138.port' and 'monitored:172.16.5.138.node_exporter_port'")
-}
-
-func (s *metaSuiteTopo) TestPlatformConflicts(c *C) {
-	// aarch64 and arm64 are equal
-	topo := Specification{}
-	err := yaml.Unmarshal([]byte(`
-global:
-  os: "linux"
-  arch: "aarch64"
-tidb_servers:
-  - host: 172.16.5.138
-    arch: "arm64"
-tikv_servers:
-  - host: 172.16.5.138
-`), &topo)
-	c.Assert(err, IsNil)
-
-	// different arch defined for the same host
-	topo = Specification{}
-	err = yaml.Unmarshal([]byte(`
-global:
-  os: "linux"
-tidb_servers:
-  - host: 172.16.5.138
-    arch: "aarch64"
-tikv_servers:
-  - host: 172.16.5.138
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "platform mismatch for '172.16.5.138' between 'tidb_servers:linux/arm64' and 'tikv_servers:linux/amd64'")
-
-	// different os defined for the same host
-	topo = Specification{}
-	err = yaml.Unmarshal([]byte(`
-global:
-  os: "linux"
-  arch: "aarch64"
-tidb_servers:
-  - host: 172.16.5.138
-    os: "darwin"
-tikv_servers:
-  - host: 172.16.5.138
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "platform mismatch for '172.16.5.138' between 'tidb_servers:darwin/arm64' and 'tikv_servers:linux/arm64'")
-}
-
-func (s *metaSuiteTopo) TestCountDir(c *C) {
-	topo := Specification{}
-
-	err := yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: "test-deploy"
-tikv_servers:
-  - host: 172.16.5.138
-    data_dir: "/test-data/data-1"
-pd_servers:
-  - host: 172.16.5.138
-    data_dir: "/test-data/data-2"
-`), &topo)
-	c.Assert(err, IsNil)
-	cnt := topo.CountDir("172.16.5.138", "/home/test1/test-deploy/pd-2379")
-	c.Assert(cnt, Equals, 2)
-	cnt = topo.CountDir("172.16.5.138", "") // the default user home
-	c.Assert(cnt, Equals, 4)
-	cnt = topo.CountDir("172.16.5.138", "/test-data/data")
-	c.Assert(cnt, Equals, 0) // should not match partial path
-
-	err = yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: "/test-deploy"
-tikv_servers:
-  - host: 172.16.5.138
-    data_dir: "/test-data/data-1"
-pd_servers:
-  - host: 172.16.5.138
-    data_dir: "/test-data/data-2"
-`), &topo)
-	c.Assert(err, IsNil)
-	cnt = topo.CountDir("172.16.5.138", "/test-deploy/pd-2379")
-	c.Assert(cnt, Equals, 2)
-	cnt = topo.CountDir("172.16.5.138", "")
-	c.Assert(cnt, Equals, 0)
-	cnt = topo.CountDir("172.16.5.138", "test-data")
-	c.Assert(cnt, Equals, 0)
-
-	err = yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: "/test-deploy"
-  data_dir: "/test-data"
-tikv_servers:
-  - host: 172.16.5.138
-    data_dir: "data-1"
-pd_servers:
-  - host: 172.16.5.138
-    data_dir: "data-2"
-  - host: 172.16.5.139
-`), &topo)
-	c.Assert(err, IsNil)
-	// if per-instance data_dir is set, the global data_dir is ignored, and if it
-	// is a relative path, it will be under the instance's deploy_dir
-	cnt = topo.CountDir("172.16.5.138", "/test-deploy/pd-2379")
-	c.Assert(cnt, Equals, 3)
-	cnt = topo.CountDir("172.16.5.138", "")
-	c.Assert(cnt, Equals, 0)
-	cnt = topo.CountDir("172.16.5.139", "/test-data")
-	c.Assert(cnt, Equals, 1)
-
-	err = yaml.Unmarshal([]byte(`
-global:
-  user: "test1"
-  ssh_port: 220
-  deploy_dir: deploy
-  data_dir: data
-tidb_servers:
-  - host: 172.16.4.190
-    imported: true
-    deploy_dir: /home/tidb/deploy
-pd_servers:
-  - host: 172.16.4.190
-    imported: true
-    name: pd_ip-172-16-4-190
-    deploy_dir: /home/tidb/deploy
-    data_dir: /home/tidb/deploy/data.pd
-    log_dir: /home/tidb/deploy/log
-`), &topo)
-	c.Assert(err, IsNil)
-	cnt = topo.CountDir("172.16.4.190", "/home/tidb/deploy")
-	c.Assert(cnt, Equals, 5)
-}
-
 func (s *metaSuiteTopo) TestGlobalConfig(c *C) {
 	topo := Specification{}
 	err := yaml.Unmarshal([]byte(`
@@ -725,69 +470,4 @@ item7 = 700
 	merge2, err := merge2Toml(ComponentTiKV, merge1, spec.TiKVServers[0].Config)
 	c.Assert(err, IsNil)
 	c.Assert(string(merge2), DeepEquals, expected)
-}
-
-func (s *metaSuiteTopo) TestTiSparkSpecValidation(c *C) {
-	topo := Specification{}
-	err := yaml.Unmarshal([]byte(`
-pd_servers:
-  - host: 172.16.5.138
-    port: 1234
-tispark_masters:
-  - host: 172.16.5.138
-    port: 1235
-tispark_workers:
-  - host: 172.16.5.138
-    port: 1236
-  - host: 172.16.5.139
-    port: 1235
-`), &topo)
-	c.Assert(err, IsNil)
-
-	topo = Specification{}
-	err = yaml.Unmarshal([]byte(`
-pd_servers:
-  - host: 172.16.5.138
-    port: 1234
-tispark_masters:
-  - host: 172.16.5.138
-    port: 1235
-  - host: 172.16.5.139
-    port: 1235
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "a TiSpark enabled cluster with more than 1 Spark master node is not supported")
-
-	topo = Specification{}
-	err = yaml.Unmarshal([]byte(`
-pd_servers:
-  - host: 172.16.5.138
-    port: 1234
-tispark_workers:
-  - host: 172.16.5.138
-    port: 1235
-  - host: 172.16.5.139
-    port: 1235
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "there must be a Spark master node if you want to use the TiSpark component")
-
-	err = yaml.Unmarshal([]byte(`
-pd_servers:
-  - host: 172.16.5.138
-    port: 1234
-tispark_masters:
-  - host: 172.16.5.138
-    port: 1236
-tispark_workers:
-  - host: 172.16.5.138
-    port: 1235
-  - host: 172.16.5.139
-    port: 1235
-  - host: 172.16.5.139
-    port: 1236
-    web_port: 8089
-`), &topo)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "the host 172.16.5.139 is duplicated: multiple TiSpark workers on the same host is not supported by Spark")
 }
